requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# File handling
python-multipart==0.0.6
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3

# AI/LLM and Vision APIs
anthropic==0.40.0
google-cloud-vision==3.7.2
openai==1.54.0
aiohttp==3.10.5

# OCR Fallback (optional - for local receipt analysis)
paddlepaddle==2.6.2
paddleocr==3.3.1

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Background Tasks
apscheduler==3.10.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# E2E Testing with Playwright
playwright==1.40.0
pytest-playwright==0.4.3
pytest-base-url==2.1.0
pytest-html==4.1.1
